+++ wazo_dird/plugins/conference_backend/http.py
@@ -59,7 +59,7 @@ class ConferenceContactList(AuthResource):
     def get(self, source_uuid):
         tenant_uuid = Tenant.autodetect().uuid
         visible_tenants = self.get_visible_tenants(tenant_uuid)
-        list_params = contact_list_param_schema.load(request.args).data
+        list_params = contact_list_param_schema.load(request.args)
         source_config = self._source_service.get(
             'conference', source_uuid, visible_tenants
         )
@@ -68,10 +68,10 @@ class ConferenceContactList(AuthResource):
         try:
             response = confd.conferences.list(**list_params)
         except Exception as e:
-            raise exception.XiVOConfdError(confd, e)
+            raise exception.WazoConfdError(confd, e)
 
         return {
             'total': response['total'],
             'filtered': response['total'],
-            'items': contact_list_schema.dump(response['items']).data,
+            'items': contact_list_schema.dump(response['items']),
         }