+++ wazo_dird/plugins/google_backend/schemas.py
@@ -2,25 +2,15 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from xivo.mallow_helpers import ListSchema as _ListSchema
-from xivo.mallow.validate import Range
-from wazo_dird.schemas import BaseSchema, BaseSourceSchema, VerifyCertificateField
+from wazo_dird.schemas import BaseSourceSchema, BaseAuthConfigSchema
 from xivo.mallow import fields
-
-from xivo.mallow.validate import Length
-
-
-class _AuthConfigSchema(BaseSchema):
-
-    host = fields.String(validate=Length(min=1, max=1024), missing='localhost')
-    port = fields.Integer(validate=Range(min=1, max=65535), missing=9497)
-    verify_certificate = VerifyCertificateField(missing=True)
-    timeout = fields.Float(validate=Range(min=0, max=3660))
-    version = fields.String(validate=Length(min=1, max=16), missing='0.1')
 
 
 class SourceSchema(BaseSourceSchema):
 
-    auth = fields.Nested(_AuthConfigSchema, missing={})
+    auth = fields.Nested(
+        BaseAuthConfigSchema, missing=lambda: BaseAuthConfigSchema().load({})
+    )
 
 
 class ListSchema(_ListSchema):