+++ wazo_dird/plugins/office365_backend/schemas.py
@@ -3,7 +3,7 @@
 
 
 from xivo.mallow_helpers import ListSchema as _ListSchema
-from wazo_dird.schemas import BaseSourceSchema
+from wazo_dird.schemas import BaseAuthConfigSchema, BaseSourceSchema
 from xivo.mallow import fields
 
 from xivo.mallow.validate import Length
@@ -11,9 +11,8 @@ from xivo.mallow.validate import Length
 
 class SourceSchema(BaseSourceSchema):
 
-    auth = fields.Dict(
-        keys=fields.String(validate=Length(min=1, max=64)),
-        values=fields.String(validate=Length(min=1, max=64)),
+    auth = fields.Nested(
+        BaseAuthConfigSchema, missing=lambda: BaseAuthConfigSchema().load({})
     )
     endpoint = fields.String(
         missing='https://graph.microsoft.com/v1.0/me/contacts',