+++ wazo_dird/plugins/profiles/schemas.py
@@ -1,6 +1,7 @@
 # Copyright 2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from marshmallow import EXCLUDE
 from xivo.mallow import fields
 from xivo.mallow.validate import Length
 from xivo.mallow_helpers import ListSchema as _ListSchema
@@ -12,7 +13,7 @@ class ResourceSchema(BaseSchema):
 
 
 class ServiceConfigSchema(BaseSchema):
-    sources = fields.Nested(ResourceSchema, many=True, missing=[])
+    sources = fields.Nested(ResourceSchema, many=True, missing=[], unknown=EXCLUDE)
     options = fields.Dict(missing={})
 
 
@@ -23,7 +24,7 @@ class ServiceDictSchema(fields.Nested):
 
         result = {}
         for service_name, service_config in nested_obj.items():
-            result[service_name] = ServiceConfigSchema().dump(service_config).data
+            result[service_name] = ServiceConfigSchema().dump(service_config)
         return result
 
     def _deserialize(self, nested_obj, attr, obj, **_):
@@ -32,7 +33,7 @@ class ServiceDictSchema(fields.Nested):
 
         result = {}
         for service_name, service_config in nested_obj.items():
-            result[service_name] = ServiceConfigSchema().load(service_config).data
+            result[service_name] = ServiceConfigSchema().load(service_config)
         return result
 
 
@@ -41,8 +42,8 @@ class ProfileSchema(BaseSchema):
     uuid = fields.UUID(dump_only=True)
     tenant_uuid = fields.UUID(dump_only=True)
     name = fields.String(validate=Length(min=1, max=512), required=True)
-    display = fields.Nested(ResourceSchema)
-    services = ServiceDictSchema(BaseSchema, required=True)
+    display = fields.Nested(ResourceSchema, unknown=EXCLUDE)
+    services = ServiceDictSchema(BaseSchema, required=True, unknown=EXCLUDE)
 
 
 class ListSchema(_ListSchema):