+++ wazo_dird/plugins/wazo_user_backend/schemas.py
@@ -1,6 +1,7 @@
 # Copyright 2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from marshmallow import EXCLUDE
 from xivo.mallow import fields
 from xivo.mallow_helpers import ListSchema as _ListSchema
 from wazo_dird.schemas import (
@@ -12,8 +13,12 @@ from wazo_dird.schemas import (
 
 
 class SourceSchema(BaseSourceSchema):
-    auth = fields.Nested(AuthConfigSchema, missing={})
-    confd = fields.Nested(ConfdConfigSchema, missing={})
+    auth = fields.Nested(
+        AuthConfigSchema, missing=lambda: AuthConfigSchema().load({}), unknown=EXCLUDE
+    )
+    confd = fields.Nested(
+        ConfdConfigSchema, missing=lambda: ConfdConfigSchema().load({}), unknown=EXCLUDE
+    )
 
 
 class ListSchema(_ListSchema):